time_bar variables are now size_t
git-svn-id: file:///home/roberto/Git/svn@7757 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: b8c5ab87a71edcae1c3ee63cab401f18f364c48c

+++ trunk/xmipp/libraries/data/program.h
@@ -245,7 +245,7 @@ protected:
     /// Use this flag for not producing a time bar
     bool allow_time_bar;
     /// Some time bar related counters
-    int time_bar_step, time_bar_size, time_bar_done;
+    size_t time_bar_step, time_bar_size, time_bar_done;
     /// Flag to know when input is a single image or stack
     bool single_image;
     bool input_is_stack;
